A number of changes to suppress unsigned vs. signed warnings

+++ main.cpp
@@ -322,7 +322,7 @@ void HandleEdge(
 
   int len = elev_a.size();
 
-  for(size_t i=0;i<len;i++){
+  for(int i=0;i<len;i++){
     auto c_l = label_a[i];
     if(c_l>1) c_l+=label_a_offset;
 
@@ -489,7 +489,7 @@ void Producer(std::vector< std::vector< ChunkInfo > > &chunks){
 
     chunks[y][x].label_offset = label_offset;
 
-    for(int l=0;l<this_job.graph.size();l++)
+    for(int l=0;l<(int)this_job.graph.size();l++)
     for(auto const &skey: this_job.graph[l]){
       label_t first_label  = l;
       label_t second_label = skey.first;
@@ -503,8 +503,8 @@ void Producer(std::vector< std::vector< ChunkInfo > > &chunks){
   }
 
   std::cerr<<"Handling adjacent edges and corners..."<<std::endl;
-  for(size_t y=0;y<gridheight;y++)
-  for(size_t x=0;x<gridwidth;x++){
+  for(int y=0;y<gridheight;y++)
+  for(int x=0;x<gridwidth;x++){
     if( (y*gridwidth+x)%10==0 )
       std::cerr<<"\tha: "<<(y*gridwidth+x)<<"/"<<(gridheight*gridwidth)<<"\n";
 
@@ -835,17 +835,17 @@ void Preparer(
 
     //nullChunks imply that the chunks around them have edges, as though they
     //are on the edge of the raster.
-    for(int y=0;y<chunks.size();y++)
-    for(int x=0;x<chunks[0].size();x++){
+    for(int y=0;y<(int)chunks.size();y++)
+    for(int x=0;x<(int)chunks[0].size();x++){
       if(chunks[y][x].nullChunk)
         continue;
       if(y-1>0 && x>0 && chunks[y-1][x].nullChunk)
         chunks[y][x].edge |= GRID_TOP;
-      if(y+1<chunks.size() && x>0 && chunks[y+1][x].nullChunk)
+      if(y+1<(int)chunks.size() && x>0 && chunks[y+1][x].nullChunk)
         chunks[y][x].edge |= GRID_BOTTOM;
       if(y>0 && x-1>0 && chunks[y][x-1].nullChunk)
         chunks[y][x].edge |= GRID_LEFT;
-      if(y>0 && x+1<chunks[0].size() && chunks[y][x+1].nullChunk)
+      if(y>0 && x+1<(int)chunks[0].size() && chunks[y][x+1].nullChunk)
         chunks[y][x].edge |= GRID_RIGHT;
     }
 
@@ -905,7 +905,7 @@ void Preparer(
       }
     }
 
-    if(retention_base=="@retainall" && chunks.size()*chunks[0].size()>=world.size()-1){
+    if(retention_base=="@retainall" && ((int)(chunks.size()*chunks[0].size()))>=world.size()-1){
       std::cerr<<"This job requires "<<(chunks.size()*chunks[0].size()+1)<<" processes. Only "<<world.size()<<" are available."<<std::endl;
       env.abort(-1);
     }
